creating include directory for GTests
