+++ src/test/CMakeLists.txt
@@ -1,2 +1,2 @@
-file(GLOB_RECURSE CPP_TEST_SOURCES ./cpp/*.cc)
+file(GLOB_RECURSE CPP_TEST_SOURCES ./cpp/src/*.cc)
 build_GenomicsDB_test_executable(runAllGTests ${CPP_TEST_SOURCES})
