+++ src/test/cpp/src/test_mapping_data_loader.cc
@@ -20,29 +20,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#include <iostream>
-#include <fstream>
-#include <getopt.h>
-#include "headers.h"
-#include "vid_mapper_sql.h"
-#include <limits.h>
-#include "gtest/gtest.h"
-
-const std::string config_file_name = "/tmp/sql_mapper/sql_mapper_config.txt";
-
-class MappingDataLoaderTester {
-  private:
-    std::vector<ContigInfo> m_contig_idx_to_info;
-    std::vector<std::pair<int64_t, int>> m_contig_begin_2_idx;
-    std::vector<std::pair<int64_t, int>> m_contig_end_2_idx;
-    std::vector<CallSetInfo> m_row_idx_to_info;
-    std::unordered_map<std::string, int64_t> m_callset_name_to_row_idx;
-  public:
-    MappingDataLoaderTester(const SQLVidMapperRequest&);
-    void validate_contig_info();
-    void validate_callset_info();
-    ~MappingDataLoaderTester();
-};
+#include <test_mapping_data_loader.h>
 
 MappingDataLoaderTester::~MappingDataLoaderTester() {
   m_contig_idx_to_info.clear();
@@ -108,60 +86,6 @@ void MappingDataLoaderTester::validate_callset_info() {
   std::cout <<"------------------------------------------------------\n";
   return;
 }
-
-/**
- * Following config parameters are expected in the file:
- * /tmp/sql_mapper/sql_mapper_config.txt
- * If the file does not exist, then default values are used
- * "host_name"  - "localhost"
- * "user_name" - "postgres"
- * "pass_word" - "postgres"
- * "db_name" - "gendb"
- * "work_space" - "/tmp/sql_mapper/workspace"
- * "array_name" - "sql_mapper_test"
- */
-class SQLMapperTest : public ::testing::Test {
-  public:
-    static MappingDataLoaderTester* loaderTester;
-
-  protected:
-    static void SetUpTestCase() {
-      std::cout <<"------------ MappingData Loader - BEGIN -----------------\n";
-      SQLVidMapperRequest request;
-      std::ifstream config_file(config_file_name);
-
-      if (config_file.is_open()) {
-        getline(config_file, request.host_name);
-        getline(config_file, request.user_name);
-        getline(config_file, request.pass_word);
-        getline(config_file, request.db_name);
-        getline(config_file, request.work_space);
-        getline(config_file, request.array_name);
-        config_file.close();
-      } else {
-        request.host_name = "localhost";
-        request.user_name = "postgres";
-        request.pass_word = "postgres";
-        request.db_name = "gendb";
-        request.work_space = "/tmp/sql_mapper/workspace";
-        request.array_name = "sql_mapper_test";
-      }
-
-      loaderTester = new MappingDataLoaderTester(request);
-      std::cout <<"------------ MappingData Loader -  END  -----------------\n";
-    }
-
-    static void TearDownTestCase() {
-        delete loaderTester;
-        loaderTester = NULL;
-    }
-
-    virtual void SetUp() {
-    }
-
-    virtual void TearDown() {
-    }
-};
 
 MappingDataLoaderTester* SQLMapperTest::loaderTester = NULL;
